refactor: rename directory
