--- lib/blockchain/README.md
@@ -1,3 +1,0 @@
-# bwallet blockchain
-
-blockchain business logic lives in here
