+++ lib/utilities/README.md
@@ -1,0 +1,4 @@
+# bwallet utilities
+
+Various utilities and helper functions live in here
+Useful things will be moved to `@bpanel/bpanel-utils`
