+++ lib/utilities/hardwareWallet.js
